swagger, querydsl 적용 user repo,entity 생성

+++ src/main/java/com/demo/domain/repository/BoardRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import java.util.Map;
 
-public interface BoardRepository extends CrudRepository<BoardEntity, Integer> {
+public interface BoardRepository extends CrudRepository<BoardEntity, Long> {
     List<BoardEntity> findAll();
 }
